Move lock.py to _internal
