+++ certbot-apache/certbot_apache/tests/configurator_test.py
@@ -73,7 +73,7 @@ class MultipleVhostsTest(util.ApacheTest):
     def test_prepare_locked(self):
         server_root = self.config.conf("server-root")
         self.config.config_test = mock.Mock()
-        os.remove(os.path.join(server_root, ".certbot.lock"))
+        os.remove(os.path.join(server_root, ".certbot._internal.lock"))
         certbot_util.lock_and_call(self._test_prepare_locked, server_root)
 
     @mock.patch("certbot_apache.parser.ApacheParser")
