+++ certbot/tests/lock_test.py
@@ -1,4 +1,4 @@
-"""Tests for certbot.lock."""
+"""Tests for certbot._internal.lock."""
 import functools
 import multiprocessing
 import unittest
@@ -17,24 +17,24 @@ from certbot.tests import util as test_util
 
 
 class LockDirTest(test_util.TempDirTestCase):
-    """Tests for certbot.lock.lock_dir."""
+    """Tests for certbot._internal.lock.lock_dir."""
     @classmethod
     def _call(cls, *args, **kwargs):
-        from certbot.lock import lock_dir
+        from certbot._internal.lock import lock_dir
         return lock_dir(*args, **kwargs)
 
     def test_it(self):
         assert_raises = functools.partial(
             self.assertRaises, errors.LockError, self._call, self.tempdir)
-        lock_path = os.path.join(self.tempdir, '.certbot.lock')
+        lock_path = os.path.join(self.tempdir, '.certbot._internal.lock')
         test_util.lock_and_call(assert_raises, lock_path)
 
 
 class LockFileTest(test_util.TempDirTestCase):
-    """Tests for certbot.lock.LockFile."""
+    """Tests for certbot._internal.lock.LockFile."""
     @classmethod
     def _call(cls, *args, **kwargs):
-        from certbot.lock import LockFile
+        from certbot._internal.lock import LockFile
         return LockFile(*args, **kwargs)
 
     def setUp(self):
@@ -93,7 +93,7 @@ class LockFileTest(test_util.TempDirTestCase):
                 os.remove(path)
             return stat(path)
 
-        with mock.patch('certbot.lock.filesystem.os.stat') as mock_stat:
+        with mock.patch('certbot._internal.lock.filesystem.os.stat') as mock_stat:
             mock_stat.side_effect = delete_and_stat
             self._call(self.lock_path)
         self.assertFalse(should_delete)
@@ -105,9 +105,9 @@ class LockFileTest(test_util.TempDirTestCase):
 
     def test_unexpected_lockf_or_locking_err(self):
         if POSIX_MODE:
-            mocked_function = 'certbot.lock.fcntl.lockf'
+            mocked_function = 'certbot._internal.lock.fcntl.lockf'
         else:
-            mocked_function = 'certbot.lock.msvcrt.locking'
+            mocked_function = 'certbot._internal.lock.msvcrt.locking'
         msg = 'hi there'
         with mock.patch(mocked_function) as mock_lock:
             mock_lock.side_effect = IOError(msg)
@@ -120,9 +120,9 @@ class LockFileTest(test_util.TempDirTestCase):
 
     def test_unexpected_os_err(self):
         if POSIX_MODE:
-            mock_function = 'certbot.lock.filesystem.os.stat'
+            mock_function = 'certbot._internal.lock.filesystem.os.stat'
         else:
-            mock_function = 'certbot.lock.msvcrt.locking'
+            mock_function = 'certbot._internal.lock.msvcrt.locking'
         # The only expected errno are ENOENT and EACCES in lock module.
         msg = 'hi there'
         with mock.patch(mock_function) as mock_os:
