+++ certbot/tests/util.py
@@ -22,7 +22,7 @@ from cryptography.hazmat.primitives import serialization
 from certbot import configuration
 from certbot import constants
 from certbot import interfaces
-from certbot import lock
+from certbot._internal import lock
 from certbot import storage
 from certbot import util
 from certbot.compat import os
