--- docs/api/lock.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.lock`
------------------------
-
-.. automodule:: certbot.lock
-   :members:
