+++ tests/lock_test.py
@@ -19,7 +19,7 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization, hashes  # type: ignore
 from cryptography.hazmat.primitives.asymmetric import rsa
 
-from certbot import lock
+from certbot._internal import lock
 from certbot import util
 from certbot.compat import filesystem
